Rework the DAS Settings Panel Layout
HowlyBlood moved the zone options into the advanced settings sub-menu, and that's about it. This resulted in activation settings being 3 levels deep, a nightmare to navigate.

I started by resorting the zone list to make it more logical, and a few hours later I arrived at a drastically different layout.

A number of options were made functional, a few LUA errors and typos were squashed in the process.

The next step will be to tokenise all this for localisation.